pyproject.toml
[dependency-groups]
dev = [
    "livekit-agents[openai,silero]>=1.3.12",
    "pydantic>=2.12.5",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "spacy>=3.8.11",
]